Use sizeof(uint32_t) etc instead of literals like 4
This is admittedly a bit of prep for supporting 16-bit bytes but I think it
can also stand on it's own as being somewhat more explicit than literals.
In the case of `const int left` it's really just using the same form as the
existing, and related, previous line does for the modulus.