Provide headers and isolate test function from main()
For use as a library rather than copy/pasting into your own codebase it is
helpful to have headers and to be able to include the tests in another
build without dealing with multiple main() functions.  For example when
cross-compiling the library for use in an embedded target and wanting to
run tests directly in the device.