Use size_t and fixed width format specifiers for printf()...  for siphash.c
It was forgotten in #17 to apply this to siphash.c as well as halfsiphash.c.  Note
that as-is this does change to a single 16-hex-char string for v0/1/2/3 in debug
output rather than having 8 then a space then 8.  If preferred it can be changed
back to having the space and using `PRIx32` twice instead of `PRIx64` once.